+++ packages/messaging/content/src/audio.ts
@@ -1,0 +1,7 @@
+import { Content } from './content'
+
+export interface AudioContent extends Content {
+  type: 'audio'
+  audio: string
+  title?: string
+}
