+++ packages/messaging/content/src/carousel.ts
@@ -1,0 +1,7 @@
+import { CardContent } from './card'
+import { Content } from './content'
+
+export interface CarouselContent extends Content {
+  type: 'carousel'
+  items: CardContent[]
+}
