+++ packages/messaging/content/src/content.ts
@@ -1,0 +1,3 @@
+export interface Content {
+  type: string
+}
