+++ packages/messaging/content/src/file.ts
@@ -1,0 +1,7 @@
+import { Content } from './content'
+
+export interface FileContent extends Content {
+  type: 'file'
+  file: string
+  title?: string
+}
