+++ packages/messaging/content/src/image.ts
@@ -1,0 +1,7 @@
+import { Content } from './content'
+
+export interface ImageContent extends Content {
+  type: 'image'
+  image: string
+  title?: string
+}
