+++ packages/messaging/content/src/location.ts
@@ -1,0 +1,9 @@
+import { Content } from './content'
+
+export interface LocationContent extends Content {
+  type: 'location'
+  latitude: number
+  longitude: number
+  address?: string
+  title?: string
+}
