+++ packages/messaging/content/src/text.ts
@@ -1,0 +1,7 @@
+import { Content } from './content'
+
+export interface TextContent extends Content {
+  type: 'text'
+  text: string
+  markdown?: boolean
+}
