+++ packages/messaging/content/src/video.ts
@@ -1,0 +1,7 @@
+import { Content } from './content'
+
+export interface VideoContent extends Content {
+  type: 'video'
+  video: string
+  title?: string
+}
