+++ packages/messaging/content/src/voice.ts
@@ -1,0 +1,6 @@
+import { Content } from './content'
+
+export interface VoiceContent extends Content {
+  type: 'voice'
+  audio: string
+}
